refactor: split error handling
  - split error types into separate modules for better maintainability
  - improve database error mapping to handle schema issues correctly
